add NewsPage folder
