+++ src/utils/router.ts
@@ -2,7 +2,7 @@ import VueRouter, { Route } from "vue-router"
 import Vue from "vue"
 import { DEFAULT_TITLE } from "@/utils/static"
 import Home from "@/components/Home/Home.vue"
-import NewsPage from "@/components/News/NewsPage.vue"
+import NewsPage from "@/components/News/NewsPage/NewsPage.vue"
 import NewsPostPage from "@/components/News/NewsPostPage/NewsPostPage.vue"
 import TeamPage from "@/components/About/TeamPage.vue"
 import HistoryPage from "@/components/About/HistoryPage.vue"
